+++ molo/surveys/wagtail_hooks.py
@@ -1,25 +1,18 @@
 from django.conf import settings
+from django.conf.urls import include, url
 from django.utils.html import format_html_join
-from django.contrib.auth.models import User
 
 from wagtail.contrib.modeladmin.options import modeladmin_register
 from wagtail.wagtailcore import hooks
 
 from molo.surveys.models import MoloSurveyPage, SurveyTermsConditions
+from molo.surveys import admin_urls
 from molo.core.models import ArticlePage
 
 from .admin import SegmentUserGroupAdmin
 
 
 modeladmin_register(SegmentUserGroupAdmin)
-
-
-@hooks.register('construct_main_menu')
-def show_surveys_entries_for_users_have_access(request, menu_items):
-    if not request.user.is_superuser and not User.objects.filter(
-            pk=request.user.pk, groups__name='Moderators').exists():
-        menu_items[:] = [
-            item for item in menu_items if item.name != 'surveys']
 
 
 @hooks.register('insert_global_admin_js')
@@ -52,3 +45,12 @@ def create_new_page_relations(request, page, new_page):
                             .first()
                         relation.terms_and_conditions = new_article
                         relation.save()
+
+
+# This overrwrites the wagtailsuveys admin urls in order to use custom
+# survey index view
+@hooks.register('register_admin_urls')
+def register_admin_urls():
+    return [
+        url(r'^surveys/', include(admin_urls)),
+    ]